# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barbearia",
        "description": "Documentação dos endpoints da barbearia desenvolvida em Laravel 11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local (Docker)"
        }
    ],
    "paths": {
...
        "/servicos": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Listar catálogo de serviços",
                "description": "Método responsável por listar todo o catálogo de serviços cadastrados na barbearia.",
                "operationId": "32212bcd3c456626b182f8e506a638c4",
                "responses": {
                    "200": {
                        "description": "Lista de serviços retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Adicionar novo serviço ao catálogo",
                "description": "Método responsável por adicionar e persistir um novo serviço ao catálogo da barbearia.",
                "operationId": "296661a8b644322875d7953720b18e8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "preco",
                                    "duracao_minutos"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Corte Social + Barba completíssimo"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 65.5
                                    },
                                    "duracao_minutos": {
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço adicionado com sucesso"
                    }
                }
            }
        },
        "/servicos/{id}": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Obter detalhes de um serviço específico",
                "description": "Método responsável por obter e exibir os detalhes de um serviço específico através do seu código identificador.",
                "operationId": "6baef081008af6eb65307d18605b2488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do serviço retornados"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Remover um serviço do catálogo",
                "description": "Método responsável por remover definitivamente um serviço do catálogo do sistema.",
                "operationId": "0b68ece74c8bf256e0e7dad496ef0a0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serviço removido com sucesso"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Atualizar preço ou dados do serviço",
                "description": "Método responsável por atualizar o preço ou outros dados cadastrais de um serviço existente no catálogo.",
                "operationId": "47cc77572cfbd122f2210a4698618e5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 70
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço updated com sucesso"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                }
            }
        }
    },
...
}